"Refactor file extensions and improve responsive design"
This commit includes several changes to improve the responsive design of the application and refactor the file extensions from .tsx to .ts where no JSX is used. The `StyledNavBarIcon` has been implemented for better style consistency. The `NavBarComponent` and `CrewCardGridComponent` have been updated for better viewing on smaller devices. The `StyledCardButton` has been replaced with a new `JoinButton` component that takes a `isFull` prop to dynamically set the button's background color and text. Other changes include adjustments to the `StyledCard`, `StyledCardInfo`, and `StyledCardImageHeader` components for better layout and readability. A typo in the `.eslintrc.cjs` file has also been corrected.