Integrated useAuthToken into useGetCrew hook
The useGetCrew hook has been modified to use the useAuthToken hook. This alteration facilitates fetching the crew details by adding token-based authentication. In the event of an error while obtaining the token or crew details, the error state is set to the encountered error and loading state is set to false.