Replace StyledCardImageHeader with ActivityCrewCardHeader
Switched out the generic StyledCardImageHeader for the new ActivityCrewCardHeader which displays an activity image based on the activity type associated with the card. This enhances the visual representation of the crew's activity in both the creation and viewing of crew cards. Other minor layout adjustments were made for better UI consistency.