Improve navigation and layout for Crew features
This commit includes modifications for the Crew features in our application. It adjusts the behavior of the 'Go Back' button to include all '/crews/' URL paths for enhanced navigation. It also changes the redirect path to the newly created Crew's page after Crew creation instead of the homepage. Finally, it optimizes the layout by removing padding in the RecentCrewCardContainer.