Add feature to update citizen name
A new feature has been implemented to allow the update of the citizen name in the player's profile. The 'updateCitizenName' action has been added in the PlayerProfileSlice, and is now available in the hook 'usePlayerProfile'. This action is dispatched in the setCitizenName method, which is now being used in the CreateProfileComponent.