Adjust input restrictions and font sizes
The minimum and maximum length for the crew description input field have been modified to allow more flexibility in the input. The font sizes for the Sign in with Discord button and the media query in the MenuButton component are also respectively decreased for better UI/UX on different screen sizes.