Update crew member links and page navigation
Crew member links have been updated to use CitizenName instead of Id. The function to join a crew has been modified to redirect a user to the crew's page upon successful joining. A UI update was made in the CrewViewWrapper component adjusting the margin. The crew join button no longer navigates a member away from the page to the home screen after joining.