Refactor CrewViewComponent and add "time ago" feature
New components, MemberList and HandleNameLink, have been created to streamline the CrewViewComponent code. The component now displays additional information such as languages and the time since the crew was created. This uses a new hook, useTimeAgo, for showing "time ago" in a human-friendly format. Fixed a typo in UpdateProfileNameComponent and improved spacing and headings in StyledCardInfo.