Add Discord invite link creation to CrewView
The revision introduces a hook 'useCreateChannelInvite', which utilizes a newly created 'DiscordServices' API to generate an invite link for a Discord channel. This is used in the CrewViewWrapper component to show the Discord invite link to the end user if a VoiceChannelId is available for the crew. The BaseAPIService postRequest function was also updated to allow an optional body parameter.