Update error handling and add profile existence check
The error handling in the App component has been revamped to print the error to the console and check if the error message includes "404". Meanwhile in the CreateProfileComponent, an additional condition has been added to the useEffect hook - it no longer only runs if the user is not logged in but also checks for the existence of profile.CitizenName. This adjustment results in redirection to the homepage if the user is not logged in or if the user has already set up a profile. In the BaseAPIService component, the error thrown now includes the response status code directly, simplifying the prior process.