Update media queries to apply styles at 775px
The CSS media query breakpoints have been updated to apply specific styles when the viewport is 775px or less instead of the previous 750px (or 400px in some cases). In addition to this, some margin-bottom changes have been added for certain styled components. This adjustment ensures that the layout and responsiveness of our components are in sync with typical device sizes.